update tema por subtema en evaluaciones y preguntas

+++ utils/api_interaccion/eliminar_archivos_vs.py
@@ -8,7 +8,7 @@ load_dotenv()
 client = OpenAI()
 
 response = client.vector_stores.files.list(
-    vector_store_id="vs_68c9c2c5a9748191b952550e1a4fbab6", limit=100)
+    vector_store_id="vs_692e087391448191b1ab0ca35298e92f")
 
 archivos = response.data
 
@@ -18,7 +18,8 @@ for a in archivos:
     print(a.attributes)
 
 for a in archivos:
-    client.vector_stores.files.delete(file_id=a.id, vector_store_id="vs_68c9c2c5a9748191b952550e1a4fbab6")
-    print(f"Archivo con ID {a.id} eliminado de la vector store")
-    client.files.delete(file_id=a.id)
-    print(f"Archivo con ID {a.id} eliminado de la API")
+    if (a.id != 'file-E4dr4KynALZWxXy8xZkwP2' and a.id != 'file-XRbssPawsGdpZQPnQx4GpW' and a.id != 'file-NoyHqcfE78eieCxtv3EvkN'):
+        client.vector_stores.files.delete(file_id=a.id, vector_store_id="vs_692e087391448191b1ab0ca35298e92f")
+        print(f"Archivo con ID {a.id} eliminado de la vector store")
+        client.files.delete(file_id=a.id)
+        print(f"Archivo con ID {a.id} eliminado de la API")
